extract PlayerComponent into AppComponent
