+++ src/app/mod.rs
@@ -192,51 +192,4 @@ impl App {
             }
         });
     }
-
-    fn player_ui(&mut self, ui: &mut egui::Ui) {
-        ui.horizontal(|ui| {
-            let stop_btn = ui.button("■");
-            let play_btn = ui.button("▶");
-            let pause_btn = ui.button("⏸");
-            let prev_btn = ui.button("|◀");
-            let next_btn = ui.button("▶|");
-
-            let mut volume = self.player.as_ref().unwrap().volume;
-            ui.add(
-                egui::Slider::new(&mut volume, (0.0 as f32)..=(5.0 as f32))
-                    .logarithmic(false)
-                    .show_value(false)
-                    .clamp_to_range(true),
-            );
-            self.player.as_mut().unwrap().set_volume(volume);
-
-            if let Some(_selected_track) = &self.player.as_mut().unwrap().selected_track {
-                if stop_btn.clicked() {
-                    self.player.as_mut().unwrap().stop();
-                }
-
-                if play_btn.clicked() {
-                    self.player.as_mut().unwrap().play();
-                }
-
-                if pause_btn.clicked() {
-                    self.player.as_mut().unwrap().pause();
-                }
-
-                if prev_btn.clicked() {
-                    self.player
-                        .as_mut()
-                        .unwrap()
-                        .previous(&self.playlists[(self.current_playlist_idx).unwrap()])
-                }
-
-                if next_btn.clicked() {
-                    self.player
-                        .as_mut()
-                        .unwrap()
-                        .next(&self.playlists[(self.current_playlist_idx).unwrap()])
-                }
-            }
-        });
-    }
 }
